MATRIX-MULTIPLICATION-WITH-PTHREAD-LIBRARY
If we increase the number of processes/threads and keep the efficiency fixed without increasing the problem size, the problem is strongly scalable. As seen in the Efficiency Table graphic, lines tend to be straight line, which means the efficiency score of this project tends to be stable so that can be said that this project is strongly scalable.